commit-change-58

+++ backend/auth/router/auth.py
@@ -1,13 +1,19 @@
 from fastapi import APIRouter,Depends,HTTPException,status,Response
 from sqlmodel import Session,select
 from core.database import get_session
-from auth.security import hash_password,verify_password,create_access_token
+from auth.security import hash_password,verify_password,create_access_token,get_current_user
 from schemas.user import UserCreate,UserOut,UserLgoin,UserUpdate,ChangePassword
 from models.user import User
 
 
 router=APIRouter(prefix="auth/",tags=["Authentication"])
 
+
+@router.get("/user", response_model=UserOut, status_code=status.HTTP_200_OK)
+def get_user(current_user: User = Depends(get_current_user)):
+    return current_user
+  
+             
 @router.post("/register",status_code=status.HTTP_201_CREATED)
 def register(user_data:UserCreate,session:Session=Depends(get_session)):
 
@@ -33,10 +39,15 @@ def register(user_data:UserCreate,session:Session=Depends(get_session)):
     return {"message": "تم إنشاء الحساب بنجاح", "user_id": new_user.id}
 
 @router.patch("/update/{CIN}")
-def update_user(CIN:str,user_data:UserUpdate,session:Session=Depends(get_session)):
+def update_user(CIN:str,user_data:UserUpdate,session:Session=Depends(get_session),current_user:User=Depends(get_current_user)):
     user=session.exe(select(User).where(User.CIN==CIN)).first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="المستخدم غير موجود")
+    if current_user.CIN != CIN:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="غير مسموح لك بتعديل بيانات مستخدم آخر!"
+        )
    
     update_data=user_data.model_dump(exclude_unset=True)
 
@@ -50,10 +61,15 @@ def update_user(CIN:str,user_data:UserUpdate,session:Session=Depends(get_session
 
 
 @router.delete("/delete/{CIN}")
-def delete_user(CIN:str,session:Session=Depends(get_session)):
+def delete_user(CIN:str,session:Session=Depends(get_session),current_user:User=Depends(get_current_user)):
     user=session.exe(select(User).where(User.CIN==CIN)).first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="المستخدم غير موجود")
+    if current_user.CIN != CIN:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="غير مسموح لك حدف بيانات مستخدم آخر!"
+        )
     session.delete(user)
     session.commit()
     return {"message":"تم حذف المستخدم بنجاح"}
@@ -78,10 +94,25 @@ def logout(response:Response):
 
 
 @router.patch("/change-passowrd/")
-def change_password(user_data:ChangePassword,session:Session=Depends(get_session)):
+def change_password(user_data:ChangePassword,session:Session=Depends(get_session),current_user:User=Depends(get_current_user)):
     
-   
-        pass
+    if not verify_password(user_data.old_password,current_user.password):
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,detail="كلمة المرور القديمة غير صحيحة")
+    
+    if not user_data.old_password ==user_data.new_password:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=""
+        )
+    hashed_new_password=hash_password(user_data.new_password)
+
+    current_user.password=hashed_new_password
+
+    session.add(current_user)
+
+    session.commit()
+
+    return {"message":"تم تغيير كلمة المرور بنجاح"}
